refactor(core): 🔨 move Metadata class to own file
